Chnaged the way props were passed to the colorslider
